update redux connection
